+++ CHANGELOG.md
@@ -14,6 +14,10 @@ Corrected mismatch in the conversion of Jeol nuclei to human
 - *processing.mask_sgn_basl*
 - *processing.apk*
 - *processing.smooth_g*
+10/10/2025
+- *processing.abca*
+- *processing.absa*
+- *Spectrum_1D.absa*
 
 ## Modified functions:
 02/10/2025
@@ -23,6 +27,8 @@ Corrected mismatch in the conversion of Jeol nuclei to human
 - *misc.trim_data_2D*: avoid truncation of one point due to slicing
 06/10/2025
 - *Spectrum_1D.qfil*: new argument **from_procs**
+10/10/2025
+- *Spectrum_1D.abs*: now returns the employed baseline
 
 
 # Version 0.4a.10